# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ExecuTorch Bencher API",
    "description": "Internal Android LLM benchmark collection and analysis service. Provides HTTP endpoints to ingest, store, and retrieve benchmark run results, content-addressed artifacts, and registered model assets.",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/artifacts": {
      "post": {
        "tags": [
          "artifacts"
        ],
        "summary": "Stream and store an artifact's content, deduplicating by SHA-256. The\nbody is the raw artifact bytes, streamed directly - **not** a multipart\nform. `kind` and `original_name` are query parameters.",
        "operationId": "uploadArtifact",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "description": "The artifact's kind. Determines whether the content is stored Zstandard-compressed (see `ArtifactKind`).",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ArtifactKind"
            }
          },
          {
            "name": "original_name",
            "in": "query",
            "description": "Caller-supplied display name, retained only as metadata.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Raw artifact bytes, streamed directly. Rejected once the stream exceeds the server's configured maximum artifact upload size.",
          "content": {
            "application/octet-stream": {
              "schema": {
                "$ref": "#/components/schemas/BinaryBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The artifact was stored, or an identical one (by SHA-256) already existed and was reused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "`kind` is not a recognized artifact kind.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The upload exceeds the server's configured maximum artifact size.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal storage error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{id}/content": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Stream an artifact's content, decompressing transparently if it is\nstored Zstandard-compressed. The response body is the raw content with\nits stored media type - never a JSON-encoded string.",
        "operationId": "getArtifactContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The artifact's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The artifact's content, streamed with its stored media type (or `application/octet-stream` if none was recorded).",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryBody"
                }
              }
            }
          },
          "404": {
            "description": "No artifact exists with this ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "The database record exists but the file is missing on disk.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{id}/download": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Like `content`, but sets `Content-Disposition: attachment` with a safe\nfilename derived from the artifact's kind and content hash.",
        "operationId": "downloadArtifact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The artifact's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The artifact's content as an attachment download.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryBody"
                }
              }
            }
          },
          "404": {
            "description": "No artifact exists with this ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "The database record exists but the file is missing on disk.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{id}/metadata": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Fetch an artifact's metadata, including whether its content-addressed\nfile is currently available on disk.",
        "operationId": "getArtifactMetadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The artifact's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The artifact's metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactMetadataResponse"
                }
              }
            }
          },
          "404": {
            "description": "No artifact exists with this ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "description": "The `error` object inside an [`ApiErrorResponse`].",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, machine-readable failure category. Current values: `invalid_request`,\n`not_found`, `artifact_file_missing`, `payload_too_large`, `conflict`,\n`not_implemented`, `internal_error`. This field is an open string, not a\nclosed enum - new categories may be added as the API grows.",
            "example": "invalid_request"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the failure. Not guaranteed stable across\nversions or releases - match on `code`, not this text.",
            "example": "gpu_clock_mhz must be greater than zero"
          },
          "details": {
            "description": "Optional structured context about the failure, e.g. which field was invalid."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque identifier correlating this error with server-side logs, when available."
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "The consistent JSON envelope every rejected or failed request responds with.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        }
      },
...
      "ArtifactKind": {
        "type": "string",
        "description": "The kinds of files the backend manages as content-addressed artifacts.\nSee `specs/benchmark-schema` - \"Large inputs and outputs are tracked as\ncontent-addressed artifacts\".",
        "enum": [
          "prompt",
          "stdout",
          "stderr",
          "output",
          "crash_log",
          "logcat",
          "correctness_report"
        ]
      },
      "ArtifactMetadataResponse": {
        "type": "object",
        "description": "An artifact's full metadata, including where to fetch its content.",
        "required": [
          "id",
          "kind",
          "size_bytes",
          "compression",
          "sha256",
          "created_at",
          "available",
          "content_url",
          "download_url"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "original_filename": {
            "type": [
              "string",
              "null"
            ]
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "media_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "compression": {
            "$ref": "#/components/schemas/Compression"
          },
          "sha256": {
            "type": "string",
            "description": "Lowercase 64-character hexadecimal SHA-256 digest of the original,\nuncompressed content.",
            "example": "b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944"
          },
          "created_at": {
            "type": "string"
          },
          "available": {
            "type": "boolean",
            "description": "Whether the content-addressed file still exists on disk. `false`\nmeans the database record exists but the file is unavailable - see\nthe `artifact_file_missing` error on `content`/`download`."
          },
          "content_url": {
            "type": "string"
          },
          "download_url": {
            "type": "string"
          }
        }
      },
...
      "BinaryBody": {
        "type": "string",
        "format": "binary",
        "description": "OpenAPI-only marker for a raw, unstructured binary body (streamed as\n`application/octet-stream`, not JSON and not a multipart form)."
      },
      "Compression": {
        "type": "string",
        "description": "How an artifact's stored bytes relate to its recorded content: `None`\nmeans the stored file *is* the original content; `Zstd` means the stored\nfile is a Zstandard-compressed encoding of it. The recorded `sha256`\nalways identifies the original, uncompressed content either way.",
        "enum": [
          "none",
          "zstd"
        ]
      },
...
      "UploadResponse": {
        "type": "object",
        "description": "Result of a successful artifact upload.",
        "required": [
          "id",
          "sha256",
          "size_bytes",
          "compression"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sha256": {
            "type": "string",
            "description": "Lowercase 64-character hexadecimal SHA-256 digest of the original,\nuncompressed content.",
            "example": "b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "compression": {
            "$ref": "#/components/schemas/Compression",
            "description": "How the stored bytes relate to the original content. `sha256` and\n`size_bytes` always describe the original, uncompressed content\neither way."
          }
        }
      },
...
    }
  },
...
}